backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
APScheduler==3.10.4
beautifulsoup4==4.12.2
selenium==4.15.2
pdfplumber==0.10.3
requests==2.31.0
sendgrid==6.11.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
lxml==4.9.3
gunicorn==21.2.0
